ブロック機能リファクタしました
